// LPS_Angular/src/app/models/Dashboard/SummaryModel.ts
export class SummaryModel {
    DEPT: string;
    SUB_DEPT: string;
    UNIT: string;
    ALL_PRODUCT_ALCO: string;
    INIT_TOTAL_LIMIT: number;
    INIT_TOTAL_CBAL: number;
    INIT_ASDATE: string;
    TOTAL_TARGET: number;
    CURRENT_TOTAL_LIMIT: number;
    CURRENT_TOTAL_CBAL: number;
    CURRENT_ASDATE: string;
    DIFF_TARGET: number;
    DIFF_INIT_LIMIT: number;
    DIFF_INIT_CBAL: number;
    VARIANCE_DATE: string;
    AVG_OUTSTANDING: number;
}